Update how API Loads Scorecard CSV File
We were using Google Directly for a bit, but it fails now like 75% of the time with timeout errors... or worse, no errors, and partial downloads that tank the import process.

An FTP account was created and this update just copies the file over the from the FTP location to be used directly on the same server, so this should resolve import issues.